vod/requirements.txt
grpcio==1.35.0
protobuf==3.14.0
grpcio-tools==1.35.0
boto3==1.26.51
opencv-python==4.7.0.68
psutil==5.9.4